Install controller shouldn't depend on MX_Controller - Fixes #262
Because we don't wants to load PHP's session extension at this time. If we
leave MX_Controler it will at some time load the CI3 session library. If we
do that while we couldn't check presence of module during installation
procedure, PHP will fail if session extension is not installed.

So we extend CI_Controller instead

Fixes #262