defer loading of modernizr script (#379)
In Firefox, the js console reports:

Layout was forced before the page was fully loaded. If stylesheets are not yet
loaded this may cause a flash of unstyled content.

This causes screen flickering when loading the page which disappears by adding
'defer' attribute to the script tag (and the error message as well).