Simplify & secure redirection to requested page after login implementation
When the user accesses a URL but is no more logged in, he is redirected to
the login page.

Until recently, we stored the requested URL using session data as well as the
content of the POST data which was reposted.

This can be a security issue (see CSRF). So now we forget about POST data and
only remember the requested URI with the query string. This is passed in the
query string of the login page URL.
(eg .../login?r_url=http.....)

The login page stores the information in 'r_url' field of the login form.
We redirect to that URL if login was successful.

Reverts e1488593fe44b7bdec3350de56278bc1e0aa1e00 and rewrites it another way
Reverts eaf6634047ec82cb7f321fa07e2376953d3a3d33