update requirement check & session_save_path
As per CI doc, it is not good practice to set sess_save_path to a world writable dir
like a tmp/temp dir. So set the value back to it default value which is NULL. When it
is NULL, it will use the value of PHP's session_save_path() which is normally set. At
least on debian it is well set to /var/lib/php/sessions.

At install time, it is checked whether the path used to store the sessions by CI is
writable. If not, it tells the user to set the value in the config.php file.

BTW, update requirement check for some extensions that weren't listed