Store POST data in session flashdata for reuse after login
If one is POSTing to a kalkun page but is no more logged in, he will be
forwared to the login screen. BTW he would lose the POST data.

This will permit to keep the POST data for use by the page that is loaded
after login.

BTW: don't put the "?" in URL if query string is empty