fix validation bypass on edit
If you Edit a user or a contact that is already in the DB having a
Phone number in the DB that is not valid (as per phone_number_validation)
JQuery validation would show in the form that the field is not valid
but still submit the form with the invalid number.

Workaround is to make the "remote" call async.

Found in a comment of
https://stackoverflow.com/q/2710548/15401262