plugins: don't fail on missing plugin files
When the plugin is already registered in the database but not installed anymore
on the filesystem, don't error out, but silently keep the plugin hidden from
the list of installed & available plugins.

When the files are back on the filesystem, they will show up and work again as
they were.