utils/upload_kalkun_deps_to_ppa.sh: don't auto-add repack suffix to kalkun packages
Also fix the older way to find repack suffix so that it removes only the
debian version, and not anything after the 1st hyphen (eg in 0.8.3~beta-1-1)

BTW, the algorithm to find repack suffix only worked if the version in
debian changelog was the same as the version of the upstream tag, or
in the devel branch. So it most time would give wrong result.