Plugins_model::get_plugins() return false only if it is not an array
$query->result() returns an empty array when the resultset of the query is
empty, which is an empty array in PHP.
Plugins_model::get_plugins() returned FALSE and interpreted an empty array as
a sign of an error in performing the query while it was actually a sign of
empty resultset (ie. no plugin installed).

This is a problem because when no plugin is installed it would return false
instead of an empty array. It caused issues when looping on the resultset.

It is expected to have an empty array when no plugin is installed.