install procedure: allow to check again as long as Gammu DB is missing
In DB Setup, when Gammu DB was missing, it was not possible to "check again"
it's presence after manual action from the user. The user only had the option
to continue to the next screen, which is not correct.

This adds the possibility to "check again" and forbids to continue further
until Gammu DB is installed.