fix sending of unicode messages
While the size of unicode SMS is max 70 chars when it is a single message,
when it is multipart, each part should be max 67 chars long (chars = utf16
code unit).

Moreover, the old way to split unicode messages didn't support some characters
like the emojis. The count of characters was not correct when using emojis