Make pgsql's user_settings table consistent with definition in mysql & sqlite
Otherwise, in pgsql there will be autoincrement, and in mysql & sqlite not.
The id_user is set programatically in Kalkun and shouldn't auto increment.
Moreover, in postrgresql, setting the value of an autoincermented field fivalue having autoincrement
will be the cause of an error in case there is a new insertion without setting a value
because the counter is not in sync anymore